+++ src/main/java/chess/domain/piece/King.java
@@ -8,7 +8,7 @@ import java.util.List;
 
 public class King extends Piece {
     private static final String SYMBOL = "K";
-    private static final double SCORE = 0;
+    private static final float SCORE = 0;
 
     public King(Team team, Position position) {
         super(team, SYMBOL, position, SCORE);
